cleanup and organize
