+++ src/colorizer/loaders/ILoader.ts
@@ -1,19 +1,8 @@
 import { Texture, DataTexture } from "three";
-import { FeatureDataType, FeatureArrayType as FeatureDataTypeArray } from "../utils/feature_utils";
-
-export type FeatureData = {
-  data: Float32Array;
-  tex: DataTexture;
-  min: number;
-  max: number;
-};
-export type TracksData = {
-  trackIds: Uint32Array;
-  times: Uint32Array;
-};
+import { FeatureDataType, FeatureArrayType } from "../types";
 
 export interface DataSource {
-  getBuffer<T extends FeatureDataType>(type: T): FeatureDataTypeArray[T];
+  getBuffer<T extends FeatureDataType>(type: T): FeatureArrayType[T];
   getTexture(type: FeatureDataType): DataTexture;
   getMin(): number;
   getMax(): number;
@@ -25,4 +14,3 @@ interface ILoader<DataType> {
 
 export interface IFrameLoader extends ILoader<Texture> {}
 export interface IFeatureLoader extends ILoader<DataSource> {}
-export interface ITracksLoader extends ILoader<TracksData> {}
