+++ src/colorizer/types.ts
@@ -1,0 +1,55 @@
+import {
+  FloatType,
+  PixelFormat,
+  PixelFormatGPU,
+  RedFormat,
+  RedIntegerFormat,
+  TextureDataType,
+  UnsignedByteType,
+  UnsignedIntType,
+} from "three";
+
+/** Available types for data loading (features, tracks, etc.), either as a CPU buffer or a GPU texture */
+export enum FeatureDataType {
+  F32,
+  U32,
+  U8,
+}
+
+/** Maps `FeatureDataType` to corresponding typed array type */
+export type FeatureArrayType = {
+  [T in FeatureDataType]: {
+    [FeatureDataType.F32]: Float32Array;
+    [FeatureDataType.U32]: Uint32Array;
+    [FeatureDataType.U8]: Uint8Array;
+  }[T];
+};
+
+type FeatureTypeSpec<T extends FeatureDataType> = {
+  arrayConstructor: { new (arr: number[]): FeatureArrayType[T] };
+  format: PixelFormat;
+  dataType: TextureDataType;
+  internalFormat: PixelFormatGPU;
+};
+
+/** Maps `FeatureDataType` to values required to create a valid texture of that type */
+export const featureTypeSpecs: { [T in FeatureDataType]: FeatureTypeSpec<T> } = {
+  [FeatureDataType.F32]: {
+    arrayConstructor: Float32Array,
+    format: RedFormat,
+    dataType: FloatType,
+    internalFormat: "R32F",
+  },
+  [FeatureDataType.U32]: {
+    arrayConstructor: Uint32Array,
+    format: RedIntegerFormat,
+    dataType: UnsignedIntType,
+    internalFormat: "R32UI",
+  },
+  [FeatureDataType.U8]: {
+    arrayConstructor: Uint8Array,
+    format: RedIntegerFormat,
+    dataType: UnsignedByteType,
+    internalFormat: "R8UI",
+  },
+};
